separated from output.go
